+++ axelor-account/src/main/java/com/axelor/apps/account/service/FixedAssetServiceImpl.java
@@ -52,7 +52,8 @@ public class FixedAssetServiceImpl implements FixedAssetService {
   @Override
   public FixedAsset generateAndcomputeLines(FixedAsset fixedAsset) {
 
-    BigDecimal depreciationValue = this.computeDepreciationValue(fixedAsset);
+    boolean isLinear = fixedAsset.getComputationMethodSelect().equals("linear");
+    BigDecimal depreciationValue = this.computeDepreciationValue(fixedAsset, isLinear);
     BigDecimal cumulativeValue = depreciationValue;
     LocalDate depreciationDate = fixedAsset.getFirstDepreciationDate();
     LocalDate acquisitionDate = fixedAsset.getAcquisitionDate();
@@ -73,7 +74,9 @@ public class FixedAssetServiceImpl implements FixedAssetService {
           fixedAsset.getGrossValue().subtract(fixedAssetLine.getCumulativeDepreciation()));
 
       fixedAsset.addFixedAssetLineListItem(fixedAssetLine);
-      if (counter == numberOfDepreciation) {
+      if ((!isLinear && counter == numberOfDepreciation)
+          || (isLinear && (isProrataTemporis && counter == numberOfDepreciation + 1)
+              || (!isProrataTemporis && counter == numberOfDepreciation))) {
         depreciationValue = fixedAssetLine.getResidualValue();
         cumulativeValue = cumulativeValue.add(depreciationValue);
         depreciationDate = depreciationDate.plusMonths(fixedAsset.getPeriodicityInMonth());
@@ -98,12 +101,13 @@ public class FixedAssetServiceImpl implements FixedAssetService {
           }
         } else {
           depreciationValue =
-              this.computeDepreciation(fixedAsset, fixedAssetLine.getResidualValue(), false);
+              this.computeDepreciation(
+                  fixedAsset, fixedAssetLine.getResidualValue(), false, isLinear);
         }
         depreciationDate = depreciationDate.plusMonths(fixedAsset.getPeriodicityInMonth());
       } else {
         depreciationValue =
-            this.computeDepreciation(fixedAsset, fixedAsset.getResidualValue(), false);
+            this.computeDepreciation(fixedAsset, fixedAsset.getResidualValue(), false, isLinear);
         depreciationDate = depreciationDate.plusMonths(fixedAsset.getPeriodicityInMonth());
       }
       depreciationValue = depreciationValue.setScale(scale, RoundingMode.HALF_EVEN);
@@ -114,9 +118,10 @@ public class FixedAssetServiceImpl implements FixedAssetService {
     return fixedAsset;
   }
 
-  private BigDecimal computeDepreciationValue(FixedAsset fixedAsset) {
+  private BigDecimal computeDepreciationValue(FixedAsset fixedAsset, boolean isLinear) {
     BigDecimal depreciationValue = BigDecimal.ZERO;
-    depreciationValue = this.computeDepreciation(fixedAsset, fixedAsset.getGrossValue(), true);
+    depreciationValue =
+        this.computeDepreciation(fixedAsset, fixedAsset.getGrossValue(), true, isLinear);
     return depreciationValue;
   }
 
@@ -136,11 +141,11 @@ public class FixedAssetServiceImpl implements FixedAssetService {
   }
 
   private BigDecimal computeDepreciation(
-      FixedAsset fixedAsset, BigDecimal residualValue, boolean isFirstYear) {
+      FixedAsset fixedAsset, BigDecimal residualValue, boolean isFirstYear, boolean isLinear) {
 
     int scale = Beans.get(AppBaseService.class).getNbDecimalDigitForUnitPrice();
     int numberOfDepreciation =
-        fixedAsset.getFixedAssetCategory().getIsProrataTemporis()
+        !isLinear && fixedAsset.getFixedAssetCategory().getIsProrataTemporis()
             ? fixedAsset.getNumberOfDepreciation() - 1
             : fixedAsset.getNumberOfDepreciation();
     float depreciationRate = numberOfDepreciation == 0 ? 0 : 1f / numberOfDepreciation * 100f;